PureMVC client using GAE service
The Flex client has been updated so that is talks to the Project web service.
A new attribute has been added to the project: department.
A new command has been added to the Python client: initialize.
The Participants list in the Flex client is still memory-only.
It is not persisted.